+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.2.0] / 2024-05-23
+### Features
+- Feature `ProductInfoUtils` to get `ProductInfo` using `msi.dll`.
 ### Fixed
 - Fix Revit not loading `.bundle` files properly when is already open. (Fix #7)
 ### Updated
 - Update `ApplicationPluginsUtils` to reverse extract bundle files.
 - Update `RevitInstallationUtils` to expose `GetRevitInstallations` with string component.
+- Update `RevitInstallationUtils` to use `ProductInfoUtils` to get `ProductInfo`.
 
 ## [1.1.2] / 2024-04-02
 ### Updated